% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{addCallback}
\alias{addCallback}
\title{Add callback to existing parameter}
\usage{
addCallback(object, type, FUN)
}
\arguments{
\item{type}{type of event - valid options are: `change`, `submit`, `click` (depending on the control type)}

\item{FUN}{callback function}
}
\description{
Add callback to existing parameter
}
