% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{buttonParam}
\alias{buttonParam}
\title{Create action button}
\usage{
buttonParam(name = paste0("button_", as.integer(runif(1) * 1e+06)),
  value = "Button", label = "", ...)
}
\description{
Create action button
}
