% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDiv.R
\name{compact}
\alias{compact}
\title{Display shiny tag without extra iframe}
\usage{
compact(tag)
}
\value{
given shiny tag decorated with rcloud-htmlwidgets attribute
}
\description{
Display shiny tag without extra iframe
}
