% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{dateInput}
\alias{dateInput}
\title{Create date input}
\usage{
dateInput(var, value = var, format = "yyyy-mm-dd")
}
\description{
Creates HTML for a clickable calender input that is passed to the param function
}
