% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocaps.R
\name{dispatchEvent}
\alias{dispatchEvent}
\title{Dispatch event handling to appropriate control's callback function}
\usage{
dispatchEvent(var_name, var_value, e)
}
\description{
Dispatch event handling to appropriate control's callback function
}
