% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDiv.R
\name{div}
\alias{div}
\title{Create div that will include parms widgets}
\usage{
div(..., byRow = FALSE)
}
\arguments{
\item{byRow}{alter widget display (T/F)}
}
\description{
Allows users to enter param widgets with html tools all in one call
}
\examples{
   div(h1("Start"),h1("Start"),
       mytags$input(v, val = 2, type = "number"),
       mytags$input(z, type = "date"),
       h1("End"))
   submit()
}
