% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.api.R
\name{.createControl}
\alias{.createControl}
\title{Creates rcloud.params.control structure}
\usage{
.createControl(label, name, group = "default", uiToRValueMapper,
  input_tag, user_callbacks = list())
}
\arguments{
\item{label}{the label to be displayed in the UI}

\item{name}{the name of the variable}

\item{group}{the control group}

\item{uiToRValueMapper}{function mapping UI values to R}

\item{input_tag}{shiny.tag to be used for the input}

\item{user_callbacks}{the list of callbacks specified by the user}
}
\value{
rcloud.params.control structure
}
\description{
Creates rcloud.params.control structure
}
