% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.dateParam}
\alias{.dateParam}
\title{Creates date-picker input control}
\usage{
.dateParam(name, label = NULL, r_class = "Date",
  tagFactory = function(...) {     inputTag <- tag("input", c(list(type =
  "date"), list(...))) }, ...)
}
\description{
Creates date-picker input control
}
