% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.api.R
\name{.paramFactory}
\alias{.paramFactory}
\title{Factory creating rcloud.params controls}
\usage{
.paramFactory(name, label, r_class, group = "default",
  tagFactory = function(...) {     do.call("tag", ...) },
  tagValueExtractor = .getValueFromTagAttribute,
  qsValueExtractor = .getSingleValueFromQueryParameter,
  nullValueProvider = function(value) {     if (is.null(value) ||
  any(is.na(value))) {         ""     }     else {         value     } },
  rToTagValueMapper = .rToUIControlDefaultValueMapper(),
  uiToRValueMapper = .uiToRValueMapper(r_class), ...)
}
\arguments{
\item{name}{the name of the variable}

\item{label}{the label for the control}

\item{r_class}{type of the variable}

\item{group}{the group the control belongs to}

\item{tagFactory}{the function returning form input shiny.tags}

\item{tagValueExtractor}{a function that extracts value specified via shiny.tag attributes}

\item{qsValueExtractor}{a function that extracts value from a query string}

\item{nullValueProvider}{a function that returns a value that should be used if default, query string nor shiny.tag specify a value}

\item{rToTagValueMapper}{a mapper that converts R value to a value accepted by shiny.tag}

\item{uiToRValueMapper}{a mapper that converts UI value to an R value}

\item{...}{parameters passed to shiny.tag, if 'callbacks' list is among them, it is removed before it is passed to tagFactory}
}
\value{
shiny.tag representing the produced parameter control
}
\description{
The factory is configurable via a number of strategies that address variability in the process of creating shiny.tags representing rcloud.params controls.
}
\details{
The returned value is a shiny.tag structure, the associated rcloud.params.control is persisted in the packages environment and is used to invoke control-specific callback functions.
}
