% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.textParam}
\alias{.textParam}
\title{Creates text input control}
\usage{
.textParam(name, label = NULL, r_class = "character",
  tagFactory = function(...) {     inputTag <- tag("input", c(list(type =
  "text"), list(...))) }, ...)
}
\description{
Creates text input control
}
