% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_widgets.R
\name{fixInputType}
\alias{fixInputType}
\title{fixInputType}
\usage{
fixInputType(param)
}
\description{
sets the 'type' to either select or specified input eg. text. 
             'input' is either select OR input
}
