% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_widgets.R
\name{param_set}
\alias{param_set}
\title{params_set}
\usage{
param_set(...)
}
\description{
params_set
}
\examples{
param_set(minimum = list(label = "Minimum", value = 10, min = 1, input = "number"),
                    maximum = list(label = "Maximum", value = 20, min = 1, input = "number"))
}
