% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{paramDiv}
\alias{paramDiv}
\title{Produces a div with additional attribute ('data-rcloud-htmlwidgets-compact')}
\usage{
paramDiv(...)
}
\arguments{
\item{...}{htmltools div function parameters}
}
\value{
shiny.tag
}
\description{
The resulting div will not be wrapped by iframe by rcloud.htmlwidgets.
}
