% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{paramSet}
\alias{paramSet}
\title{Creates param set form}
\usage{
paramSet(..., name = paste0("form_", as.integer(runif(1) * 1e+06)),
  wait.if.invalid = TRUE, hide.source = FALSE)
}
\arguments{
\item{...}{child elements (shiny.tags)}

\item{name}{of the form}

\item{wait.if.invalid}{should notebook execution be stopped if parameter values are invalid, if `true` and `on.submit` function is not provided, an error will be produced if form is invalid}

\item{hide.source}{should the source of the cell displaying the form be hidden}

\item{on.submit}{callback function to invoke when form is submitted}

\item{on.change}{callback function to invoke on each parameter of the param set change}
}
\value{
rcloud.params.param.set structure
}
\description{
Note! paramSet is not a shiny.tag, this means that it may not be wrapped in htmltools shiny.tag
}
