% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{print.rcloud.params.param.set}
\alias{print.rcloud.params.param.set}
\title{Prints a param set form}
\usage{
\method{print}{rcloud.params.param.set}(x, ..., view = interactive())
}
\description{
Param set can have two modes:
* reactive - it doesn't perform busy wait, it halts notebook execution and passes control over execution to JS.
* synchronous (blocking) - it performs busy wait, and R keeps control over the execution process.
}
\details{
Whether to wait or not for a form is controlled by `waitIfInvalid` logical property of the `paramSet`. 
In case of reactive form, if the flag is set to `true` (default) and `on.submit` callback
is not defined, the form will be validated and an error produced if the form is invalid.
}
