% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{rcloud.hide.source.current.cell}
\alias{rcloud.hide.source.current.cell}
\title{Hide currently executed cell's source}
\usage{
rcloud.hide.source.current.cell()
}
\description{
Hide currently executed cell's source
}
