% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{rcloud.stop.execution}
\alias{rcloud.stop.execution}
\title{Gracefully stops execution of a notebook by allowing current cell to complete}
\usage{
rcloud.stop.execution()
}
\description{
Gracefully stops execution of a notebook by allowing current cell to complete
}
