% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{rcloud.ui.append}
\alias{rcloud.ui.append}
\title{Append content to an HTML element}
\usage{
rcloud.ui.append(selector, what)
}
\arguments{
\item{selector}{target element selector, e.g. '#my-div'}

\item{what}{content to add, e.g. shiny.tag}
}
\value{
always TRUE
}
\description{
Append content to an HTML element
}
