% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{rcloud.ui.plot}
\alias{rcloud.ui.plot}
\title{Generate plot in an HTML element}
\usage{
rcloud.ui.plot(selector, plot.fun, width = 300, height = 300)
}
\arguments{
\item{selector}{target element selector, e.g. '#my-div'}

\item{plot.fun}{function producing a plot}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\value{
TRUE
}
\description{
Generate plot in an HTML element
}
