% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{rcloud.ui.prepend}
\alias{rcloud.ui.prepend}
\title{Prepends content to an HTML element}
\usage{
rcloud.ui.prepend(selector, what)
}
\arguments{
\item{selector}{target element selector, e.g. '#my-div'}

\item{what}{content to prepend, e.g. shiny.tag}
}
\value{
always TRUE
}
\description{
Prepends content to an HTML element
}
