% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{rcloud.ui.set}
\alias{rcloud.ui.set}
\title{Sets content of an HTML element}
\usage{
rcloud.ui.set(selector, what)
}
\arguments{
\item{selector}{target element selector, e.g. '#my-div'}

\item{what}{content to set, e.g. shiny.tag}
}
\value{
TRUE
}
\description{
Sets content of an HTML element
}
