% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{removeCallbacks}
\alias{removeCallbacks}
\title{Remove callbacks from existing parameter}
\usage{
removeCallbacks(object, type)
}
\arguments{
\item{type}{type of event - valid options are: `change`, `submit`, `click` (depending on the control type)}
}
\description{
Remove callbacks from existing parameter
}
