% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{submit}
\alias{submit}
\title{Returns output of widget to R side when clicked}
\usage{
submit(f = NULL)
}
\arguments{
\item{f}{optional. Run a user defined function}
}
\description{
Returns output of widget to R side when clicked
}
