% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{submitParam}
\alias{submitParam}
\title{Create submit button}
\usage{
submitParam(name = paste0("submit_", as.integer(runif(1) * 1e+06)),
  value = "Submit", label = "", ...)
}
\description{
Create submit button
}
