% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{waitForGroup}
\alias{waitForGroup}
\title{Blocks execution and waits for submission of a form associated with the given group}
\usage{
waitForGroup(group = "default")
}
\description{
TODO Consider removing this
}
