% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{waitForSynchronousForm}
\alias{waitForSynchronousForm}
\title{Blocks execution and waits for submission of a form associated with the given group}
\usage{
waitForSynchronousForm(form_id)
}
\description{
Blocks execution and waits for submission of a form associated with the given group
}
