% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.choiceParam}
\alias{.choiceParam}
\title{Creates radio buttons group input control}
\usage{
.choiceParam(name, label = NULL, choices = list(), group = "default",
  ...)
}
\description{
Creates radio buttons group input control
}
