% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.api.R
\name{.createControl}
\alias{.createControl}
\title{Creates rcloud.params.control structure}
\usage{
.createControl(label, var.name, ui.to.r.value.mapper, input.tag,
  user.callbacks = list(), par.name = var.name)
}
\arguments{
\item{label}{the label to be displayed in the UI}

\item{var.name}{the name of the variable}

\item{ui.to.r.value.mapper}{function mapping UI values to R}

\item{input.tag}{shiny.tag to be used for the input}

\item{user.callbacks}{the list of callbacks specified by the user}

\item{par.name}{parameter name, as used by the UI}
}
\value{
rcloud.params.control structure
}
\description{
Creates rcloud.params.control structure
}
