% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.api.R
\name{.createControl}
\alias{.createControl}
\title{Creates rcloud.params.control structure}
\usage{
.createControl(label, var_name, uiToRValueMapper, input_tag,
  user_callbacks = list(), par_name = var_name)
}
\arguments{
\item{label}{the label to be displayed in the UI}

\item{var_name}{the name of the variable}

\item{uiToRValueMapper}{function mapping UI values to R}

\item{input_tag}{shiny.tag to be used for the input}

\item{user_callbacks}{the list of callbacks specified by the user}

\item{par_name}{parameter name, as used by the UI}
}
\value{
rcloud.params.control structure
}
\description{
Creates rcloud.params.control structure
}
