% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.logicalParam}
\alias{.logicalParam}
\title{Creates checkbox input control}
\usage{
.logicalParam(name, label = NULL, r.class = "logical",
  tag.factory = function(...) {     tags$input(type = "checkbox", class =
  "checkbox", ...) }, empty.value.handler = .erroringEmptyValueHandler,
  tag.value.extractor = function(inputTag) {     tagValue <- NULL     if
  ("checked" \%in\% names(inputTag$attribs)) {         if
  (!is.null(inputTag$attribs$checked) &&
  !any(is.na(inputTag$attribs$checked))) {             tagValue <- if
  (is.logical(inputTag$attribs$checked)) {                
  inputTag$attribs$checked             }             else {                
  TRUE             }         }     }     tagValue },
  r.to.tag.value.mapper = .rToUIControlCheckboxValueMapper(), ...)
}
\description{
Creates checkbox input control
}
