% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.numericParam}
\alias{.numericParam}
\title{Creates numeric input control}
\usage{
.numericParam(name, label = NULL, min = NA, max = NA,
  r.class = "numeric", tag.factory = function(...) {     paramsIn <-
  list(...)     type <- "number"     if ("type" \%in\% names(paramsIn)) {  
        type <- paramsIn$type         paramsIn$type <- NULL     }    
  inputTag <- tag("input", c(list(type = type), paramsIn)) }, ...)
}
\description{
Creates numeric input control
}
