% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.numericParam}
\alias{.numericParam}
\title{Creates numeric input control}
\usage{
.numericParam(name, label = NULL, min = NA, max = NA,
  group = "default", ...)
}
\description{
Creates numeric input control
}
