% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{.selectParam}
\alias{.selectParam}
\title{Creates select input control}
\usage{
.selectParam(name, label = NULL, choices = list(),
  r.class = "character", tag.factory = function(...) {    
  tag("select", list(...)) },
  tag.value.extractor = .getValueFromTagAttribute,
  qs.value.extractor = .getMultiValueFromQueryParameter,
  empty.value.handler = function(var, val) {     c() },
  r.to.tag.value.mapper = .rToUIControlSelectValueMapper(choices), ...)
}
\description{
Creates select input control
}
