% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcloud.params.R
\name{param}
\alias{param}
\title{Pass parameters to javascript}
\usage{
param(inputTag, name, varClass)
}
\arguments{
\item{inputTag}{HTML string to create widget}

\item{name}{varibale name}

\item{varClass}{class of variable}
}
\description{
Takes HTML string and passes to javascript. Variables can be updated through widget then passed back to R
}
