% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDiv.R
\name{paramSet}
\alias{paramSet}
\title{Creates param set form}
\usage{
paramSet(..., callbacks = list(), wait_for = FALSE,
  name = paste0("form_", as.integer(runif(1) * 1e+06)),
  group = "default")
}
\arguments{
\item{...}{child elements (shiny.tags)}

\item{callbacks}{form callbacks}

\item{wait_for}{should the R process be blocked when form is displayed}

\item{name}{name of the form}

\item{group}{the group of the controls}
}
\value{
rcloud.params.param.set structure
}
\description{
Note! paramSet is not a shiny.tag, this means that it may not be wrapped in htmltools shiny.tag
}
